+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Remove `TLS.Domain` field and CLI argument.
+
 ## [0.2.0] - 2023-11-24
 
 ### Added