+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `WithError` logs error to the canonical log line.
+
+### Changed
+
+- `BadRequestWithError` handles `context.Canceled` and `context.DeadlineExceeded`
+  specially as well, like `InternalServerErrorWithError`.
+
+### Removed
+
+- `NotFoundWithError`: call `WithError` first and then `NotFound`.
+
 ## [0.4.0] - 2023-11-28
 
 ### Added