+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `APICors` and `GetCors` on `Route` struct enable CORS handling.
+
 ## [0.6.0] - 2023-12-07
 
 ### Added