+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.0] - 2024-02-16
+
 ### Added
 
 - `Router.Get` resolves path and method to a route descriptor `ResolvedRoute`,
@@ -90,7 +92,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First public release.
 
-[unreleased]: https://gitlab.com/tozd/waf/-/compare/v0.7.0...main
+[unreleased]: https://gitlab.com/tozd/waf/-/compare/v0.8.0...main
+[0.8.0]: https://gitlab.com/tozd/waf/-/compare/v0.7.0...v0.8.0
 [0.7.0]: https://gitlab.com/tozd/waf/-/compare/v0.6.0...v0.7.0
 [0.6.0]: https://gitlab.com/tozd/waf/-/compare/v0.5.0...v0.6.0
 [0.5.0]: https://gitlab.com/tozd/waf/-/compare/v0.4.0...v0.5.0